fix(scripts): retarget apps/api/scripts imports for moved files
Codex review on #972 caught that apps/api/scripts/assessments.ts still
imported ../src/assessment-cli.js and ../src/assessment-resolution.js after
this PR moves both to community/ -- every assess:* CLI command would abort
with ERR_MODULE_NOT_FOUND. apps/api/scripts is outside tsconfig's include
(rootDir/include is just "src"), so neither tsc nor CI's typecheck step
catches a broken script import; these only fail at tsx runtime.

While fixing it, found and fixed the same class of bug already live on
master from #971: apps/api/scripts/beta-invite.ts still imports
../src/mailer.js and ../src/email-templates.js from before that move.
